web/requirements.txt
flask
pymongo
opencv-python
numpy
deepface
requests
python-dotenv
reportlab
